pyproject.toml
[tool.poetry]
name = "certification"
version = "0.1.0"
description = ""
authors = ["Alexandr Zibinin <zibinin.chel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
black = "^24.10.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
